+++ Sources/KFinder/FinderController.swift
@@ -4,19 +4,20 @@ import Foundation
 final class FinderController {
     func currentFinderWindowDirectories() throws -> [URL] {
         let script = """
-        tell application "Finder"
-            set output to ""
-            repeat with finderWindow in Finder windows
-                try
-                    set output to output & POSIX path of (target of finderWindow as alias) & linefeed
-                end try
-            end repeat
-            return output
-        end tell
-        """
+            tell application "Finder"
+                set output to ""
+                repeat with finderWindow in Finder windows
+                    try
+                        set output to output & POSIX path of (target of finderWindow as alias) & linefeed
+                    end try
+                end repeat
+                return output
+            end tell
+            """
         let descriptor = try execute(script)
         let output = descriptor.stringValue ?? ""
-        return output
+        return
+            output
             .split(separator: "\n")
             .map(String.init)
             .map { URL(fileURLWithPath: $0) }